Enable boot splash on desktops
Note that Debian 9's PrisonPC bootstrap had this:

  chroot $t plymouth-set-default-theme --rebuild-initrd prisonpc

  overlay/usr/share/plymouth/themes/prisonpc:
  -rw-r--r--   17K  logo.png
  -rw-r--r--   206  prisonpc.plymouth
  -rw-r--r--  2.3K  prisonpc.script
  -rw-r--r--   173  progress-bar.png
  -rw-r--r--   195  progress-box.png

We simplify that to use a built-in theme engine (two-step), with
a built-in logo stored on the mainboard (bgrt).
This matches the style of Windows 10 boot splash.

Instead of sitting through an extra rebuild, I reverse-engineered
/usr/sbin/plymouth-set-default-theme and determined it was simply
setting Theme= in /etc/plymouth/plymouthd.conf.

So we can just set that directly.
This is the third thing that needs force-confold, so
move that up to ALWAYS be set.

Size Time
351M 214.1970s   template=desktop  (before this commit, no splash)
359M 216.1018s   template=desktop  (this commit, splash)

PS: in qemu the boot is too fast for splash to even appear.
To force a delay for testing purposes, you can do this:

    debian-11-desktop/slow-down-boot-to-test-plymouth:#!/bin/sh
    debian-11-desktop/slow-down-boot-to-test-plymouth:[ prereqs = "$1" ] && echo plymouth && exit
    debian-11-desktop/slow-down-boot-to-test-plymouth:sleep 10
    debian-11-desktop/slow-down-boot-to-test-plymouth.tarinfo:{"name": "usr/share/initramfs-tools/scripts/init-premount/slooooow", "mode": 320}