Install office suite (libreoffice)
Size Time
302M 168.1795s   virtual-only local-only template=desktop  (before this commit, no office)
412M 233.7692s   virtual-only local-only template=desktop  libreoffice
407M 214.2959s   virtual-only local-only template=desktop  libreoffice-writer libreoffice-calc libreoffice-impress (no base)
407M 212.0879s   virtual-only local-only template=desktop  libreoffice-writer libreoffice-calc libreoffice-impress libreoffice-math (no base)

When installing "libreoffice", libreoffice-base is installed but DOES NOT WORK.
You can try to create a database, but it errors:

    The connection to the external data source could not be established.
    No SDBC driver was found for the URL 'sdb:embedded:hsqldb'.
    ./connectivity/source/commontools/dbexception.cxx:399

This is because it needs libreoffice-sdbc-hsqldb, which
has a hard dependency on a Java runtime environment.
So libreoffice-base remains useless in a "java is banned" environment.

Should we just ship useless lobase ANYWAY, purely to keep the build script simpler?

I strongly suspect as soon as we try to cherry-pick from backports, this issue will become moot.