new template: DBAN (Darik's Boot and Nuke), via nwipe
DBAN exists for a specific use case.

  1. If it's an SSD, use the ATA "secure erase" opcode.
     DBAN is wrong for this, because there is an FTL in the way.

  2. If you are throwing the disk out, use physical destruction.
     A quick and cheap "good enough" method is to drill several holes through the platters.
     DBAN takes ~24 hours for a 4TB NAS drive.  Fuck that.

  3. If you were smart and used full-disk encryption (e.g. LUKS),
     just delete the decrypt keys / passwords.
     DBAN is a waste of time.

  4. You're giving your old HDD to someone you don't trust,
     AND there were secrets on it,
     AND it's spinning rust (not an SSD),
     AND you didn't use FDE (e.g. LUKS).
     DBAN is a sensible choice here.

Now the next problem is: DBAN was bought out in 2012 and shut down in 2015.
So instead you need to run "nwipe" which is a fork of the "active ingredient" of DBAN.

It's in Debian.  You just need a way to run it that (hopefully) won't
accidentally have you nwipe your laptop/server's real HDDs by accident
because you weren't paying attention.

WARNING: You DO NOT WANT --method=dodshort (the default).
         It is only effective for ancient MFM drives.
         Instead use --method=prng, which is generic.
         One pass (the default) is probably OK, I guess.

REMINDER: if you boot with "toram", you can unplug the DBAN USB key.
          This means you won't accidentally erase the DBAN USB key, either.

Size  Time
144M  121.0260s   template=main (before this commit)
149M  123.7481s   template=dban (this commit)