new template: zfs install/rescue image
https://openzfs.github.io/openzfs-docs/Getting%20Started/Debian/

When ZFS 0.7 or 0.8 was current, I could just grab Ubuntu 20.04 *desktop* ISO, and
use that to install or rescue Debian 10 buster systems with root-on-ZFS.

This avoided needing to recompile zfs-dkms in the live environment every single boot,
which was both slow and needed a lot of RAM.

Then, Debian got ZFS 2.0, while Ubuntu Live was still on 0.8.
To get the native encryption and zstd compression benefits,
I *had to* be running ZFS 2.0 *BEFORE* writing the root filesytem.

That meant Ubuntu 20.04 couldn't be used, and I needed something else.
This is/was that something else.

This might be moot now that Ubuntu has had 20.10 and 21.04 releases, but
I can't even be arsed to check.  I gave up on Ubuntu years ago.

NOTE: we don't need zfs-initramfs here, because the *live* system isn't booting off ZFS.

Size  Time
81M    72.2448s   virtual-only local-only template=main
81M    72.2290s   virtual-only local-only template=dban
150M  215.4663s   virtual-only local-only template=zfs