(UNTESTED) new template: datasafe3 (rsync pull receiver)
From 2008 to 2013 we sold an rsync+ZFS backup service remarkably
similar to what https://rsync.net/ sells today.  It was not
successful, and in 2018 we shut it down and migrated remaining
customers to a stand-alone "minimum viable product".

This is that product.

It has a single remaining deployment:
daisy.cyber.com.au, backing up
<something>.abrahall.id.au.

This is pretty much a straight copy of the bootstrap.git datasafe3
branch over to (publicly-visible) bootstrap2020.
I confirmed that it builds and boots, but THAT'S ALL!

I'm doing this much mainly so my "rsnapshot best practice" config is
in a public repo.

FIXME: turn stand-alone journalcheck into a .deb, and --include that,
       instead of having an embedded copy.

See also:

 * https://github.com/cyberitsolutions/cyber-zfs-backup (replacement)
 * https://github.com/cyberitsolutions/journalcheck (spinoff)
 * https://github.com/cyberitsolutions/datasafe-lite (orthogonal)
 * https://KB.cyber.com.au/Datasafe3
 * ssh://login.cyber.com.au/srv/vcs/zbm.git (predecessor)
 * ssh://login.cyber.com.au/srv/vcs/zbm-2017-05-22.git