Compress faster
Requires mmdebstrap 0.8, which
requires apt 2.3.7, which
requires libstdc++6 11.2.0-10, which
requires Debian 12 bookworm.

IOW this feature prevents us building on Debian 11, and
can't easily be worked around.

If you try to run this with mmdebstrap 0.7.5,
you get this misleading error:

    I: creating tarball...
    I: main() received signal PIPE: waiting for setup...
    W: creating tarball failed: E: cannot copy to standard output: Broken
    pipe

    ^CI: main() received signal INT: waiting for setup...
    E: tar failed: 2

Initial demonstration shows with zstd,
new optimize=size is faster than
old optimize=speed!

    SIZE   SPEED
     85M    74.6179s   virtual-only local-only                (before this commit, xz)
    103M   113.7662s   virtual-only local-only optimize=speed (before this commit, xz)
    103M    57.0340s   virtual-only local-only                (this commit, zstd)
    171M    27.5852s   virtual-only local-only optimize=speed (this commit, lz4)

This is also interesting as an upstream-maintained alternative to delete-bad-files.
The only things missing are:

  • mmtarfilter doesn't log WHICH files get excluded (makes debugging harder), and
  • mmtarfilter lacks a --from-file option so each path-exclude=X line
    must be passed as a CLI argument.