PrisonPC: add repo for in-house kludge and non-free packages
The server side of this lives at ssh://login.cyber.com.au/srv/vcs/cyber-ansible.git:cyber-apt-ftparchive-generate.
We need a full repo so that apt can see it right away.
If we just did something like

    --customize-hook=chroot $1 wget https://github.com/X/Y/releases/Z.deb
    --customize-hook=chroot $1 apt install ./Z.deb

Then the package Z is trying to workaround would already be installed.
In some cases that is fine;
in other cases it makes a mess that is fiddly to clean up.

PS: the reason path validation now allows A-Z is because
I use ~/Desktop/ as my 'dumping ground' / 'scratch area', and
I can't be arsed moving that.
Likewise it felt bad to work around it by doing

    shutil.copyfile('foo.asc', td / 'foo.asc')

so that then I could tell mmdebstrap an absolute path relative to the
already-validated tempfile.TemporaryDirectory().