desktop: watch store-bought DVD movies
You can see the compiler costs quite a bit in space, but not time.

Initially I was going to just install libdvdss2 in the non-PrisonPC base desktop image, but
that would immediately make it unbuildable for non-CyberIT people.

I also considered just not including DVD support in the base desktop.
In the end, because the built TIME wasn't affected too much,
I decided to use upstream's method and accept a little waste.

(Caveat: this is timing with --comp zstd, but really these commits
apply --comp xz, which magnifies time cost somewhat.)

SIZE TIME
552M 165.3496s template=desktop (before)
623M 172.4282s template=desktop (after, libdvd-pkg + build-deps)
553M 163.0337s template=desktop-inmate (before)
553M 164.2113s template=desktop-inmate (after, libdvdcss2 prebuilt)