inmate: explicitly half-and-catch-fire if PrisonPC vlc is out of date
What we want:

    1) if PrisonPC's vlc is newest, use it;
    2) otherwise, abort and ask for human help.

In Debian 9 our strategy was roughly:

    apt install {every vlc binary package name}=1.2.3-1inmate1
    ...
    apt list --upgradable | grep "up to date, nothing to do"

If Debian vlc was newest, it would be detected as a pending upgrade,
and the second line would fail grep, aborting the build at the end.

This was a bit annoying for a couple of reasons:

    a. we had to name each dependency explicitly.

       This was quite verbose.

    b. we had to give the version number explicitly.

       This meant a plinking git commit every time.

We still have the equivalent of that apt|grep command,
it is called

    debian-11-main.hooks/customize90-abort-on-unexpected-upgrade.py

HOWEVER, because we now do src:vlc pinning, and
just --include=vlc (solving #1 and #a and #b),
our previous implicit global defense will not catch this (#2).

This is important enough that we can just add an explicit check.
Additional benefits are:

  1. as an essential hook, this now fires much earlier.
     I am less likely to have already wandered off to get a coffee.

  2. as an specific explicit check, it can suggest a specific explicit fix.