popcon: do not generate WM_CLASS_NAME2menu lookup table until we need it
Earlier the loop was done INSIDE this script, so it generated the table once, then used it.
Now the loop runs way outside ourselves, and (oops) we are re-generating this lookup table.
Initially, just move it down so it is skipped when no window is active.
Later, I guess we want to do cache it?  It really depends how slow the lookup table is to generate.
Honestly, it probably is not a big deal.