what-is-my-ip: go back to xdg autostart
Without xfce being properly systemd integrated (no graphical-session-pre.target &c),
I cannot see how to ensure systemd starts what-is-my-ip AFTER /etc/X11/Xsession has told systemd what the DISPLAY and XAUTHORITY are.
Without this, what-is-my-ip will start too early, crash, and that is the end of it.

I could have used a .timer with something like OnActiveSec=10, so it just starts the .service 10s after login.
That is a shitty hack and -- for now -- xdg autostart is a less-shitty hack.

I hate this.