upload-to: try to sidestep brain damage on re-upload
This was the error I saw:

    '/srv/netboot/images/desktop-staff-latest' -> '/srv/netboot/images/desktop-staff-previous'
    cp: '/srv/netboot/images/desktop-staff-previous/site.dir' and '/srv/netboot/images/desktop-staff-2022-01-18/site.dir' are the same file

The upload code really isn't as robust as I thought it was, I guess...
Needs someone to brain all the way through it while fully awake.
I can't manage that short-term, so sidestep the worst case, which is when X-latest and X-previous are identical.