add dummy DVD to tests
I used a mini.iso for now because it's small (50MB) and contains a valid filesystem.
My other thoughts were to download/make a small test music CDDA or movie DVD, or
to use genisoimage to make a "hello world.txt" ISO9660 (not UDF, so a bit suck).

Since mini.iso is hosted on Debian, I used http:// and the existing acng proxy, so
it is only downloaded once.

I was going to use arm64 instead so it wouldn't be directly usable, but
that image is 90MB instead of 50MB.