disc snitch: use correct URL for the first time since 2014
As far back as 2014, we had

    curl -fd @- https://trimserver.lan/discokay <<<"host_ip=$host_ip&uid=$active_uid&label=$ID_FS_LABEL&summary=$disc_report_summary" )

and in the same year,

    ['curl', '-sLSf', 'https://ppc-services/discokay',

AFAICT that hostname change to ppc-services has ALWAYS been wrong.
On the master server, https://ppc-services/ --> pete, and https://PrisonPC/ --> eric.
And eric is the only thing that implements /discokay.

So this SHOULD have been failing for ever, but
we can definitely see that it is WORKING right now with the Debian 9 images using curl https://ppc-services/discokay.

I cannot explain why it is working when it should not be.

However, in the boot-test VM, we DO need to use the correct URL.
This is most likely because either

  * the Debian 11 images use python3's urllib (not curl); or

  * ppc-services is an A record in /etc/hosts,
    whereas on real hardware,
    ppc-services is a CNAME to prisonpc