rsyslog: read /*/log/journal/*.journal (not /run/systemd/journal/syslog)
If systemd-journald provided a RELP push instead of HTTP pull,
we would not need any of this bullshit.

The rsyslog docs say basically:

  1. journald was created by Lennart in secret with no consultation
     from existing logging experts (even personal friends like Gerhard),
     and is a bit dumb.

  2. imjournal can break when .journal files are corrupt (due to journald bugs).
     When this happens, rsyslogd will logspam downstream log servers.
     Therefore do not use imjournal unless you really have to.

  3. imuxsock defaults to /dev/log and will delete anything already
     there, put a ryslog listener there, and remove it when rsyslog
     exists.

     BUT if rsyslog detects that systemd is around,
     it will instead use /run/systemd/journal/syslog.
     This is why journald is hard-coded to WRITE to.

     This can also break due to journald bugs, but
     when this happens, the result is lost messages, not log spam.
     Therefore use this one.

We have used imuxsock just fine with a hard-coded rsyslog.conf for a decade.

We recently switched to using a systemd unit that configures
rsyslog.conf dynamically for the local network, based on SRV records.
When we did that, rsyslog broke and started listening on /dev/log
instead of / as well as /run/systemd/journal/syslog.

The problem GOES AWAY if we turn off all get-config-from-dnssd* files.
i.e. SOMEHOW we caused this.

I tried for about 6 hours to isolate and fix this problem, by
juggling the systemd unit hooks.  In the end I gave up.

I am using imjournal instead now which I think works better???
I'm slightly scared of #1 though.

https://www.rsyslog.com/doc/v8-stable/configuration/modules/imuxsock.html#coexistence-with-systemd
https://www.rsyslog.com/doc/v8-stable/configuration/modules/imjournal.html