delete-bad-files: bugfix: A/__pycache__/B.cpython-39.pyc (not A/B.pyc)
The old globs written for python2 were assuming "import A.B" meant A/B.py and A/B.pyc.
But after PEP 3147, they are installed in a runtime-specific path.

This was further confused by
https://github.com/cyberitsolutions/bootstrap2020/blob/main/debian-11-PrisonPC.hooks/customize95-obfuscate-python-inner.py
which reverts PEP 3147 by the time you boot.
But it runs AFTER delete-bad-files, so the globs need to look for the PEP 3147 path.