delete-bad-files: migrate diagnostic binary shitlists
As a rough test for "is this program ever called", I did this from a
booted template=desktop-inmate:

    root@desktop-inmate:~# find / -xdev -type f -exec grep --color -nHFw systemd-analyze {} +

Where "systemd-analyze" is the program being deleted.
In general, it only shows up inside the program itself (for --help output, I guess).

"ldconfig" had a script called during glib's postinst.
I removed the script as well, because fuck it.

"ldd" and "shred" had a few false positives.