Force niceness
I thought about this for a long time.
What tipped the balance is, the AMC image 12GB to 13GB of /tmp during the build.
I have 16GB total RAM.  So I resizes /tmp to 13GB, queued a build, and went to bed.
When I got up, my entire desktop was gone, because with "only" 3GB left, gnome3 had triggered the OOM killer.
Now, space in /tmp is not counted against RAM waste by the OOM killer.
Until I can think of a way to do that, at least mark this process as a batch job.
That will implicitly tell most OOM killers that this is a higher priority for killing.

I could also directly set oom_score_adj, but nice does that implicitly and does not really hurt.
I think this makes xz compression take slightly longer on an 8-core laptop, but... meh.