Always abort on --debug
I thought about this for a long time.
What tipped the balance for me is this:
if you cleanly exit from --debug, something weird happens which causes mmdebstrap to background itself.
If you then foreground it and hit ^C to interrupt the built, our wrapper script leaves some crap around in /tmp wasting space.
(That last problem can/should be fixed by not hard-coding /tmp/bootstrap2020 as the default destdir, but instead using a tempfile.TemporaryDirectory when --remove is passed.)

Anyway, the use case of "I want to use --debug to tweak something, then continue to build and boot-test" is not something I have ever really needed in practice, so fuck it.
(It was much more useful in Debian 9 era when bootstrap.git took 20 to 40 minutes, instead of 2 to 5 minutes, for a full regular build.)

The line removing /etc/debian_chroot file could be removed, but I have left it in for now, because if this change gets reverted, it would not immediately be obvious that debian_chroot was leftover mess.