dvdrip: remove needless magic number
The point of "progessfunc 0.005" was so that as soon as ripping starts,
the progress bar changes from "pulse back and forth" to "0% finished".

This is not really interesting, because we can just trust vlc to update its state promptly-ish.
And if it doesn't because vlc broke, well, leaving it as a throbber is not LYING.