fix Lodge Complaint w/ Debian 9 PrisonPC main server
This is also fixed in Debian 11 PrisonPC main server.
Fixing EITHER end is sufficient.

OK, eric's https://prisonpc/Complain works from w3m on the server itself
But not in chromium...
eric[467381]: TypeError: main() got multiple values for argument 'complaint_text'
OK I found the fucking prolem
If you browse to https://prisonpc/Complain it works.  If you browse to https://prisonpc/Complain/ it fails
Because action= doesn't expect that
And "Complain" (no trailing slash) is done in eric/eric-links.cfg, but
"https://prisonpc/Complain/" is in the SOE:
https://github.com/cyberitsolutions/bootstrap2020/blob/d8a5bdeb9be6b51b2dbe92364174561fd40abd0b/debian-11-PrisonPC/chromium/01-Startup.json#L42-L43
So that's an easy fix at least.