desktop: add "compress PDF" to thunar Send To menu
https://alloc.cyber.com.au/task/task.php?taskID=34659

The use case is lawyers who run their photocopier/scanner in default 1200DPI mode,
which makes 100MB+ PDFs, when the inmate really only needs a lower resolution (say 5MB).

We do not want to provide /usr/bin/gs because that allows password-locking PDFs.
Inmates tried opening PDFs in lodraw, but we block fine-grained export-to-PDF for the same reason.
All prisons that ship gimp already have libgs9 installed, so we can add a simple,
narrow-focused way to address this issue.
We can do so very easily because we do not even need to add a dedicated GUI, we can reuse existing Send To menu in the file manager.

Value adds: use libnotify to make a 'I am done' popup appear?
            use exo-open to open the output PDF afterwards, so the user can verify that the compressed image is still readable?
            have libnotify say something like 'new file is X% smaller' and if it is NOT smaller, just delete it.

            maybe have more than just 96DPI and 300DPI?
            I am not sure what numbers are 'reasonable' here.