desktop: bring back desktop icons
> Desktop icons are missing

I can put the desktop icons back easily, but
I cannot stop the popup happening every time you double-click on a
launcher (screenshot attached).
This popup bug happened in old SOEs, too.
"No desktop icons" sidestepped the popup issue, and is the current GUI
design trend anyway.

It's no big deal either way; a major upgrade was a
good opportunity to gauge if the response is/was
"this is awful, fix it!" or "oh, I didn't even notice".

Based on the reaction so far, I think turning desktop icons back on is
the way to go :-)