Use pipewire (not pulseaudio)
This fixes a persistent bug where if you tried to change volume while vlc was playing a video, sometimes (often!) pulseaudio would just crash.
pulseaudio would restart, but the already-open vlc would never make noises again.
To fix it you had to close and reopen vlc.

    pulseaudio:
    orc_code_region_allocate_codemem():
    Failed to create write and exec mmap regions.
    This is probably because SELinux execmem check is enabled (good) and $TMPDIR and $HOME are mounted noexec (bad).

Mike found that this is because Debian pipewire is compiled with --enable-orc.
That is https://github.com/GStreamer/orc and what it does is create new executables at runtime like
/tmp/orcexec.XXXXXX or /run/user/1000/orcexec.XXXXXX or ~/orcexec.XXXXXX
This is obviously banned on PrisonPC -- all user-writable areas are mounted -o noexec precisely to defeat this class of attack.

We probably could have solved this by recompiling pulseaudio with ./configure --disable-orc.

    https://sources.debian.org/src/pulseaudio/14.2-2/debian/control/#L26

However, we already had pipewire ready to go, and
we planned to switch to pipewire permanently in Debian 12.

The only reason to stick with pulseaudio for Debian 11 was that this commit is long and yukky.
But given that it fixes and actual user-visible and SUPER ANNOYING bug, fuck it, pipewire TODAY.

PS: pipewire does not use orc at all; it's not even in the codebase.