EPG: allow you to go back from "TV/ABC" to "TV", using the mouse
With --app, it was a hiding the address bar, including the back button.
And the IPTV "app" isn't done as a "Single Page Application" so you need the back button.
SAD FACE

This also brings it back to consistency with the Debian 9 SOEs.
It was different because we went through a "chrome://apps phase".