tvserver: disable EVERY unused driver that triggers "error: too many arguments to function 'usb_maxpacket'"
This started happening in the last week or so.
It makes the TV server SOE unbuildable.
The last successful build was

    tvserver-2022-10-21-1666311758
    5.18.0-0.deb11.4-amd64 (debian-kernel@lists.debian.org) #1 SMP PREEMPT_DYNAMIC Debian 5.18.16-1~bpo11+1 (2022-08-12)

The failing builds are trying to build 5.19.
Most likely the out-of-tree TBS drivers need to be updated to cope with 5.19 kernel.

AFAICT this is happening for basically every TV Remote Control (RC_*) that is wired up to the USB bus, i.e. basically all of them.
We don't use TV remotes in our headless TV tuner server.
I don't know if the out-of-tree SA716X driver has a mandatory TV remote component.
I guess I'm going to find out...