dbus-broker not dbus
It's a different implementation, and it's more hardened.
We do not need to explicitly block dbus being installed with e.g. a Conflicts.
If both are installed, dbus-broker will "win" due to Install.Alias=dbus.service.

We do not need the machine-id symlink anymore, because
that is/was a dbus-specific backcompat hack for non-systemd systems.
dbus-broker is systemd-only so it doesn't have that hack in the first place.

https://lists.debian.org/debian-devel/2023/07/msg00117.html