Drop support for --virtual-only desktops
These were only used in testing anyway, and
the mouse pointer was invisible when using them, so
except when interacting with the VM via VNC,
it was largely a useless code path.

Instead from now on you will have to build virtual-and-physical images to test,
and put up with the extra 30 seconds of build time each time.