Fix paths that are wrong due to usrmerge
Even though dpkg -L and apt-file say the file is installed as /lib/X/Y,
when find runs, it is found at /usr/lib/X/Y.
If we write the exact path /lib/X/Y it does not match.
I could write the exact path /usr/lib/X/Y but I am likely to see that and "fix" it later,
so for the avoidance of future confusion, I am using a glob **/lib consistently now.