Use exo-open (not xdg-open → exo-open)
If you build with --no-apps, and chromium is not installed, then
exo-open (exo-utils) is installed, but
xdg-open (xdg-utils) is not installed.

Since we KNOW the desktop will always be XFCE,
we might as well just skip the needless xdg-open layer, and
call exo-open directly.

The alternative would be to add include=xdg-utils to the .tarinfo files.

Note that this is really kind of moot, because when --no-apps skips installing the GUI browser,
these 2 scripts that use xdg-open will not have the GUI browser installed to run it.

The only concrete differences are:

  1. when testing these scripts with --no-apps, the error is a little clearer.
  2. when testing with --no-apps, there is 50kB less download (ABSOLUTELY negligible).
  3. in production, these scripts start SLIGHTLY faster for end users.