Revert "Allow installing scratch without x-terminal-emulator"
This reverts commit e030a76b450bffa8134add29390270533151191b.

There is NO POINT installing Scratch 1.4 from 2011.
It is over a decade old!

Upstream also has Scratch 2.0 which is based on Adobe Flash.
Upstream also has Scratch 3.5~ which is based on Electron (i.e. chrome + nodejs).

  → https://downloads.scratch.mit.edu/desktop/Scratch%20Setup.exe

Upstream does not provide a Linux version, but a third-party group provide one:

    https://scratch.mit.edu/discuss/topic/326943/
  → https://scratux-revived.github.io/
  → https://github.com/scratux-revived/scratux/releases/download/3.29.1/scratux_3.29.1_amd64.deb

This is including an entire copy of Chrome as well as a 160MB uninspectable bundle, all in /opt.
This is, in my opinion, not a good idea and not easily fixed.

I think we should instead consider allowing some of the online version that runs in the existing chromium browser.

Otherwise we CAN ship 1.4 but I am 99.999% confident that any educator
will give them workbooks intended for 3.x and they just Will Not Work.