UKI: port to Debian 12
Building both traditional (vmlinuz + initrd.img) and UKI (LINUXX64.EFI) boot material.
The --boot-test stuff only uses the latter.
The former is deprecated but still needed (as at 2023) for
https://git.cyber.com.au/prisonpc/23.10.0_rc1 and
https://git.cyber.com.au/prisonpc/22.09.1

This commit is a dependency BEFORE prisonpc-core can have this:
https://git.cyber.com.au/prisonpc/34923-http-bootserver

Also note: *NONE* of the test desktops in Cyber IT's lab support EFI.
In theory they could be reflashed to support that, but
we probably can't have the firmware images anymore...