Sort Chromium policies alphabetically (consistent with upstream 115+)
Also put them in a separate dir in the source, because the amount of other crap in there was getting a bit much.

Keep 00-Proxy and 00-URLs with their prefixes, because those are the "important to *us*" files.
Rename 00-URLs to 10-URLs, so that you can easily say "just the proxy settings" with

    rm /etc/chromium/policies/managed/[!0]*

instead of needing to do

    rm /etc/chromium/policies/managed/{[!0]*,0[!0]*,00-URLs.json}

...which has been a source of minor annoyance for me for years.

See also https://github.com/cyberitsolutions/bootstrap2020/commit/1edb804918522edda85243c1e85c3d7c608aaf54