Automate editing db/dbx keyrings?
I now understand how to

 * use ukify genkey (or GnuTLS certtool) to generate an RSA-2048 key and certificate.
 * use python3-virt-firmware virt-fw-vars to take an "empty" OVMF_VARS_4M.fd and say "trust my new key, and ONLY my new key"
 * boot a UKI signed with that key
 * find out exactly what's in the trusted list (on VMs; still not sure about physical hardware).

I *think* I know how to say
"here's the MS master key, add it to the explicit denylist", but
I *also* think that doesn't actually work the way you expect, because
I think UEFI Secure Boot doesn't implement all of X.509, so
you have to deny closer to the edge.

I need to test that some more, but OTOH we can[REST OF SENTENCE LOST]

NEXT STEP: glue the proof-of-concept test code into existing stuff (stymied a little because some of this is only in Debian 13 / 12bpo).