POC using VM desktop, but real PrisonPC main server
Running on tweak so that it uses tcb.py and apache and the shitty in-house cert from there, as-is.

Initially it fails because linuxx64.efi served by tcb.py is not signed by the nonce cert.

Then I do this WHILE QEMU IS RUNNING:

    cyber@tweak:~$ sudo cp --verbose --backup /tmp/mmdebstrap.*//linuxx64.efi /srv/netboot/images/main-latest/linuxx64.efi

Then in qemu, do ^Ac sendkey ctrl-alt-delete ^J ^Ac

Then it loads the signed linuxx64.efi as shown attached.
Because there is no SSH port forwarding, and
there is no console=ttyS0 in the boot args,
that's as far as I can see in this test.