build-misc: cope with both "foo-1/main.c" and "./foo-1/main.c"
Almost all source tarballs look like this:

    drwxr-xr-x 1000/100          0 2020-11-09 16:03 quota-4.06/
    -rw-r--r-- 1000/100       8182 2020-11-09 16:03 quota-4.06/rquota.c
    -rw-r--r-- 1000/100      35252 2019-11-05 15:12 quota-4.06/warnquota.c
    -rw-r--r-- 1000/100      10749 2020-10-19 15:38 quota-4.06/quotaon.c

But prayer is old and hand-made and looks like this:

    drwxr-xr-x dpc22/dpc22       0 2012-07-05 09:59 ./prayer-1.3.5/
    drwxr-xr-x dpc22/dpc22       0 2012-07-05 09:59 ./prayer-1.3.5/servers/
    -rw-r--r-- dpc22/dpc22    1157 2008-09-16 09:59 ./prayer-1.3.5/servers/prayer_shared.h
    -rw-r--r-- dpc22/dpc22     424 2008-09-16 09:59 ./prayer-1.3.5/servers/session_exchange.h

GNU tar --strip-components=1 doesn't "just ignore" the leading "./".
Add a transform to do so, as a quick hack.

Note that pretty much all modern codebases do "git archive" or equivalent, not
"make distclean && make dist", so this kind of upstream inconsistency should be very rare in future.