root_squash all NFS shares (not just rw shares)
Remove security downgrade workaround for pre-rsync satellites

The shares were originally no_root_squash so that site.dir could be accessed directly by satellites.
Since then,

    * satellites switched to copying via rsync:// instead of NFS.
    * rsync had its own equivalent of root_squash, so
    * site.dir is pre-compressed into site.squashfs - no NFS clients needs to read site.dir
    * satellites themselves were removed completely

Since this share is read-only, this security enhancement is not really important.
But it is easier to always root_squash, rather than try to justify NOT doing so.

Note that for *NEW* deployments, the initial sharenfs= is set in the bootstrap2020 repo
and then may be amended by doc/Build.rst.

https://alloc.cyber.com.au/task/task.php?taskID=35286