squash! Update per code review
Move the deb dependency into the .py.tarinfo.
That way it's "next to" the import in the .py.

This makes it harder for the "apt install X" and "import X" to get out of sync, when something gets overhauled in a few years.

(The toml where this was before is more for things that *don't have* a script using them directly, so they can't get out-of-sync.)