Adblock is wrongly allowing chrome://settings
Does patching the "core" list (10-URLs.json) work better?
I *think* so, but I'm not super confident.

Is the actual issue that there's a limit of 1000 rules for URLAllowlist and URLBlocklist?
Maybe 1000 of the ~70,000 rules are effectively picked at random, and
'*' is not one of them, but only sometimes?

Also the documentation says that URLAllowlist beats URLBlocklist.
So then why does this block https://en.wikipedia.org/wiki/Lorem_ipsum ?:

    {"URLAllowlist": ["http://*", "https://*"],
     "URLBlocklist": ["*", "wikipedia.org"]}