Install uBOL only for staff, and via Google Store.
I have not actually seen this work, but *in theory* it will work.

This will only work if PrisonPC allows staff access to the store.

It seems likely this is

    0 10.0.0.1 TCP_DENIED/407 5703 POST http://update.googleapis.com/service/update2/json?cup2key=15:WpxibHgg-lEzbobvi_OGyj9R2sNfUhjmaR0nPOatwsE&cup2hreq=f93886bc60f60ff0904d28d120ce6246144863a7cc76ad1bff45b648ed5689a0 - HIER_NONE/- text/html app.id:- ETag:-

Note that staff have full access to all sites *anyway*, but they have to authenticate to squid first (using their LDAP credentials).

The backend service within chromium that downloads webexts from the store PROBABLY knows to use the proxy, but doesn't know how to authenticate.

The way we hacked around this elsewhere was to go into the squid config and say

    allow if staff and domain is deb.debian.org            <--- add googleapis here???
    allow if ldap authenticated and staff                  <--- full staff access
    allow if ldap authenticated and <complicated rules>    <--- limited detainee access

Note that we CANNOT allow only specific parts of google (e.g. specific plugins) in squid.
We can only enforce that in chromium.