rsbsign: first draft
This is also the first file in my idea of moving all the .py and
auxiliary files into a "python module" stack, so that

    ruff .
    mypy .

will Just Work.
Currently mypy gets cranky about having >1 file with the same name.

I expect to do the D13 work in there, and
probably not bother porting the D11/D12 stuff over (at least initially).

I expect to use importlib.resources for .files and .hooks,
turning the former directly into a tarball (as now), and
turning the latter into a tempdir to feed to mmdebstrap --hooks-dir.